split sensors2graph
